refactored converter
